# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.


    **Configuration environment variables:**
    `INSTANCE_ID` (auto-generated UUID; included in all API response bodies and
    webhook events), `HTTP_ADDR` (:8080), `SIP_BIND_IP`, `SIP_PORT`, `SIP_HOST`,
    `ICE_SERVERS`, `RECORDING_DIR`, `LOG_LEVEL`, `WEBHOOK_URL`,
    `ELEVENLABS_API_KEY`, `VAPI_API_KEY` (VAPI agent provider API key),
    `S3_BUCKET`, `S3_REGION`, `S3_ENDPOINT`, `S3_PREFIX`.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs:
    post:
      operationId: createLeg
      summary: Originate an outbound SIP call
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLegRequest'
      responses:
        '201':
          description: Leg created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '400':
          description: Invalid JSON, bad SIP URI, unknown codec, or unsupported type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listLegs
      summary: List all active legs
      tags: [Legs]
      responses:
        '200':
          description: Array of legs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Leg'
  /legs/{id}:
    parameters:
      - $ref: '#/components/parameters/LegId'

    get:
      operationId: getLeg
      summary: Get a single leg
      tags: [Legs]
      responses:
        '200':
          description: Leg details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '404':
          $ref: '#/components/responses/LegNotFound'

    delete:
      operationId: deleteLeg
      summary: Hang up a leg
      tags: [Legs]
      responses:
        '200':
          description: Leg hung up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: hung_up
        '404':
          $ref: '#/components/responses/LegNotFound'
  /legs/{id}/answer:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: answerLeg
      summary: Answer a ringing or early-media inbound SIP leg
      tags: [Legs]
      responses:
        '200':
          description: Answer initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: answering
        '400':
          description: Not a SIP inbound leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in ringing or early_media state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/early-media:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: earlyMediaLeg
      summary: Enable early media on a ringing inbound SIP leg
      description: >
        Sends SIP 183 Session Progress with SDP and sets up the media pipeline.
        The leg transitions to `early_media` state, allowing audio playback and
        room participation before the call is answered.
      tags: [Legs]
      responses:
        '200':
          description: Early media enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: early_media
        '400':
          description: Not a SIP inbound leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in ringing state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Media setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/mute:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: muteLeg
      summary: Mute a leg
      description: >
        A muted leg's audio is excluded from the room mix and speaking events
        are suppressed. Taps (recording/STT) still receive the muted leg's own audio.
      tags: [Legs]
      responses:
        '200':
          description: Leg muted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: muted
        '404':
          $ref: '#/components/responses/LegNotFound'

    delete:
      operationId: unmuteLeg
      summary: Unmute a leg
      tags: [Legs]
      responses:
        '200':
          description: Leg unmuted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: unmuted
        '404':
          $ref: '#/components/responses/LegNotFound'
  /legs/{id}/hold:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: holdLeg
      summary: Put a SIP call on hold
      description: >
        Sends a re-INVITE with `sendonly` SDP direction. The RTP timeout is
        paused while held, and a 2-hour auto-hangup timer starts.
      tags: [Legs]
      responses:
        '200':
          description: Call held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: held
        '400':
          description: Not a SIP leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in connected state, or already held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: unholdLeg
      summary: Resume a held SIP call
      description: Sends a re-INVITE with `sendrecv` SDP direction.
      tags: [Legs]
      responses:
        '200':
          description: Call resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: resumed
        '400':
          description: Not a SIP leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
...
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Leg:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        leg_id:
          type: string
          description: Unique leg identifier (UUID)
        type:
          type: string
          enum: [sip_inbound, sip_outbound, webrtc]
        state:
          type: string
          enum: [ringing, early_media, connected, held, hung_up]
        room_id:
          type: string
          description: Room ID if the leg is in a room, empty otherwise
        muted:
          type: boolean
          description: Whether the leg is muted
        held:
          type: boolean
          description: Whether the call is on hold (SIP legs only)
        sip_headers:
          type: object
          additionalProperties:
            type: string
          description: 'X-* headers from the inbound INVITE. Only present on sip_inbound legs.'
      required: [leg_id, type, state]
...
    Error:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
    CreateLegRequest:
      type: object
      properties:
        type:
          type: string
          enum: [sip]
          description: Leg type
        uri:
          type: string
          description: SIP URI to dial
        from:
          type: string
          description: 'Caller ID — sets the user part of the SIP From header (e.g. "+15551234567", "alice")'
        privacy:
          type: string
          description: SIP Privacy header value (e.g. "id", "none")
        ring_timeout:
          type: integer
          description: Seconds to wait for answer; 0 = no timeout
          default: 0
        max_duration:
          type: integer
          description: Maximum call duration in seconds after connect. Automatically hung up when reached. 0 or omitted = no limit.
          default: 0
        codecs:
          type: array
          items:
            type: string
            enum: [PCMU, PCMA, G722, opus]
          description: Codec preference order
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom SIP headers to include in the outbound INVITE (e.g. X-Correlation-ID)
        auth:
          type: object
          description: SIP digest authentication credentials. If the remote challenges with 401/407, sipgo will retry with these credentials.
          properties:
            username:
              type: string
              description: SIP auth username
            password:
              type: string
              description: SIP auth password
        room_id:
          type: string
          description: Room ID to auto-add the leg to once media is ready (early_media or connected). If the room does not exist, it is automatically created.
        webhook_url:
          type: string
          format: uri
          description: Route all events for this leg exclusively to this URL instead of global webhooks.
        webhook_secret:
          type: string
          description: HMAC-SHA256 signing secret for the per-leg webhook.
      required: [type, uri]
...